+++ rust/src/physics/single_chain/fjc/mod.rs
@@ -1,19 +1,17 @@
 pub mod test;
 pub mod mechanics;
 pub mod thermodynamics;
-
 pub struct FJC
 {
     pub hinge_mass: f64,
     pub link_length: f64,
-    pub number_of_links: u16,
+    pub number_of_links: u8,
     pub mechanics: mechanics::Mechanics,
     pub thermodynamics: thermodynamics::Thermodynamics
 }
-
 impl FJC
 {
-    pub fn init(number_of_links: u16, link_length: f64, hinge_mass: f64) -> FJC
+    pub fn init(number_of_links: u8, link_length: f64, hinge_mass: f64) -> FJC
     {
         FJC
         {
